fix 'make flash' wait serial sync issue
'make flash' will call 'wm_tools' to download the image to the target.
the configuration in the makefile use at command to reset the target,
witch is not working in the current version of the dev-board. the tool
will wait for the target to sync forever. this patch fix that by change
the reset action from 'at' to 'rts', using the serial port rts pin to
control the device reset.

file changed:
    - ./tools/w800/rules.mk
    - ./.gitignore